backend/requirements.txt
django>=5.1.4,<5.2
psycopg2-binary>=2.9.10,<2.10
djangorestframework>=3.15.2,<3.16